+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Change GE Wine version from GE-Proton7-31 to GE-Proton7-43.
 - Change the screenshot program from Shutter to Spectacle.
 - Add support for the Steam Deck controller.
+- Add 'pacman-static' for more stable upgrades.
 
 ## 3.2.1 - The SteamOS 3.4 Update